requirements.txt

# Backend
fastapi
uvicorn
python-multipart
pytesseract
opencv-python
pandas
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow